// src/core/modules/collection/type.d.ts
export type Collection = {
  bodyHtml: string;
  collectionType: string;
  createdDate: string;
  deleted: boolean;
  disjunctive: boolean;
  handle: string;
  image: any;
  imageSrc: any;
  published: boolean;
  rules: any[];
  sortOrder: string;
  title: string;
  updatedDate: string;
  _id: string;
};